Modify db connection
