+++ .vscode/settings.json
@@ -1,6 +1,13 @@
 {
   "editor.tabSize": 2,
-  "cSpell.words": ["coderabbit", "noopener", "noreferrer", "peaceiris", "VITE"],
+  "cSpell.words": [
+    "coderabbit",
+    "noopener",
+    "noreferrer",
+    "peaceiris",
+    "usecase",
+    "VITE"
+  ],
   "editor.defaultFormatter": "esbenp.prettier-vscode",
   "editor.formatOnSave": true,
   "togglequotes.chars": ["\"", "'", "`"],
