+++ backend/domain/book/entity.go
@@ -1,0 +1,5 @@
+package book
+
+type BookRepository interface {
+	GetBooks(userId int, shelfId string, startIndex int, maxResult int) ([]Book, error)
+}
