+++ backend/domain/book/repository_interface.go
@@ -1,0 +1,8 @@
+package book
+
+// book entity
+type Book struct {
+	ID     string
+	Title  string
+	Author string
+}
